Fix typos and make sure the secret is in the my-secrets-ch namespace
Without this change the secret never gets propagated to the other
clusters. While we're at it, let's fix a couple of typos in the README
and also s/secrets/secret/ so kustomize won't error out on this example.

With this change I was able to see the secret propagate to the other
clusters. Without it the secret gets created in the default namespace
and never gets picked up by the subscription.